requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.12.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
install==1.3.3
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
Pillow==7.2.0
pipenv==2020.8.13
psycopg2==2.8.6
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.51
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.2.0
